refactor(server): extract startup and port-recovery helpers
Move PID-file and EADDRINUSE recovery logic from server.ts into startup.ts, preserve readPpid export for tests, and remove onSend payload any-cast. Refs #350
